+++ JobTracker/tracker/forms.py
@@ -12,3 +12,14 @@ class JobPostingForm(forms.ModelForm):
         model = JobPosting
         fields = ['company_name', 'job_title', 'posting_url', 'location', 'is_remote', 'description']  # Include 'is_remote'
 
+from .models import Document
+
+class DocumentUploadForm(forms.ModelForm):
+    class Meta:
+        model = Document
+        fields = ['document_type', 'file', 'text_content']
+        widgets = {
+            'document_type': forms.Select(attrs={'class': 'form-control'}),
+            'file': forms.ClearableFileInput(attrs={'class': 'form-control'}),
+            'text_content': forms.Textarea(attrs={'class': 'form-control', 'rows': 4}),
+        }